+++ docs/trace/extending-the-sdk/Program.cs
@@ -16,6 +16,7 @@
 
 using System.Diagnostics;
 using OpenTelemetry;
+using OpenTelemetry.Trace;
 
 public class Program
 {