+++ src/OpenTelemetry.Exporter.Jaeger/JaegerExporter.cs
@@ -26,7 +26,7 @@ using Process = OpenTelemetry.Exporter.Jaeger.Implementation.Process;
 
 namespace OpenTelemetry.Exporter.Jaeger
 {
-    public class JaegerExporter : BaseExporter<Activity>
+    internal class JaegerExporter : BaseExporter<Activity>
     {
         private readonly int maxPayloadSizeInBytes;
         private readonly TProtocolFactory protocolFactory;
@@ -110,9 +110,6 @@ namespace OpenTelemetry.Exporter.Jaeger
                             continue;
                         case Resource.ServiceNamespaceKey:
                             serviceNamespace = strVal;
-                            continue;
-                        case Resource.LibraryNameKey:
-                        case Resource.LibraryVersionKey:
                             continue;
                     }
                 }