+++ src/OpenTelemetry.Exporter.Zipkin/ZipkinExporterOptions.cs
@@ -15,6 +15,7 @@
 // </copyright>
 
 using System;
+using System.Diagnostics;
 
 namespace OpenTelemetry.Exporter.Zipkin
 {
@@ -30,7 +31,8 @@ namespace OpenTelemetry.Exporter.Zipkin
 #endif
 
         /// <summary>
-        /// Gets or sets the name of the service reporting telemetry.
+        /// Gets or sets the name of the service reporting telemetry. If the `Resource` associated with the telemetry
+        /// has "service.name" defined, then it'll be preferred over this option.
         /// </summary>
         public string ServiceName { get; set; } = DefaultServiceName;
 
@@ -51,5 +53,15 @@ namespace OpenTelemetry.Exporter.Zipkin
         /// </summary>
         public int? MaxPayloadSizeInBytes { get; set; } = DefaultMaxPayloadSizeInBytes;
 #endif
+
+        /// <summary>
+        /// Gets or sets the export processor type to be used with Zipkin Exporter.
+        /// </summary>
+        public ExportProcessorType ExportProcessorType { get; set; } = ExportProcessorType.Batch;
+
+        /// <summary>
+        /// Gets or sets the BatchExportProcessor options. Ignored unless ExportProcessorType is BatchExporter.
+        /// </summary>
+        public BatchExportProcessorOptions<Activity> BatchExportProcessorOptions { get; set; } = new BatchExportProcessorOptions<Activity>();
     }
 }